+++ src/tui/layout_minimal.rs
@@ -74,7 +74,7 @@ impl LayoutMinimal {
 }
 
 fn get_block_height(len: usize, area: Rect) -> usize {
-    (len + 2).clamp(3, (area.height as f64 * 0.5).ceil() as usize)
+    (len + 2).clamp(1, (area.height as f64 * 0.5).ceil() as usize)
 }
 
 // Vertically Centered:
@@ -84,3 +84,4 @@ fn get_block_height(len: usize, area: Rect) -> usize {
 //   Constraint::Length(1),
 //   Constraint::Length(widget_h),
 //   Constraint::Fill(1),
+//